requirements.txt
flask==3.0.0
requests==2.31.0
python-dotenv==1.0.0
openai==1.3.0
playwright==1.40.0
pandas==2.1.3
pillow==10.1.0
PyPDF2==3.0.1
beautifulsoup4==4.12.2
numpy==1.26.2
